refactor: rename api.ts to queries.ts and add mutations.ts
Split entity API layer into separate files for better TanStack Query alignment:
- Renamed api.ts to queries.ts for read operations (GET)
- Added mutations.ts placeholder for write operations (POST, PUT, DELETE)
- Updated imports and documentation
