+++ app/_entities/pet/mutations.ts
@@ -1,0 +1,4 @@
+import axios from 'axios';
+
+// Pet mutation functions (create, update, delete operations)
+// Add mutation functions here as needed
